Revert "feat: 체널툭 추가 (#10)"
This reverts commit f6c9a90777f2f4a35844589f1dc40ef8dc02d635.
