--- components/Providers/index.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { ThemeProvider } from "./ThemeProvider";
-import { ChannelTalkProvider } from "./ChannelTalk/ChannelTalkProvider";
-
-interface ProvidersProps {
-  children: React.ReactNode;
-}
-
-export default function Providers({ children }: ProvidersProps) {
-  return (
-    <ThemeProvider>
-      <ChannelTalkProvider>{children}</ChannelTalkProvider>
-    </ThemeProvider>
-  );
-}
